+++ src/vuex/plugins/session-storage.js
@@ -1,11 +1,13 @@
 import { vuexTypes } from '@/vuex/types'
 import config from '@/config'
+import { ErrorTracker } from '@/js/helpers/error-tracker'
 
 export const sessionStoragePlugin = store => {
   store.subscribe((mutation, state) => {
     switch (mutation.type) {
       case vuexTypes.CLEAR_STATE: {
         localStorage.removeItem(config.STORAGE_KEY)
+        ErrorTracker.setLoggedInUser({})
         break
       }
       case vuexTypes.POP_STATE: {
@@ -26,6 +28,16 @@ export const sessionStoragePlugin = store => {
           idleHandler: savedStore.idleHandler,
         })
 
+        break
+      }
+      case vuexTypes.SET_ACCOUNT:
+      case vuexTypes.SET_WALLET: {
+        if (vuexTypes.walletAccountId && vuexTypes.walletEmail) {
+          ErrorTracker.setLoggedInUser({
+            'accountId': vuexTypes.walletAccountId,
+            'email': vuexTypes.walletEmail,
+          })
+        }
         break
       }
       default: